Create README.md
Hi! This is a simple Node.js + Express API I built for the Mediversal technical assignment. The goal of this project was to create an endpoint that lets users upload product images. Those images are stored in AWS S3, and the image URLs are saved in a MySQL database hosted on AWS RDS. The whole thing is deployed using Docker on an EC2 instance.